compilesketches: use GITHUB_SHA to get commit hash
This commit changes the way the HEAD commit hash is retrieved in the
'get_head_commit_hash' function. Instead of using 'git.Repo' to retrieve
the hash, it now directly uses the 'GITHUB_SHA' environment variable.
This avoids the need to have a full Git repository in the current
working directory when the action runs in clean contexts (such as the
Zephyr test actions).

Signed-off-by: Luca Burelli <l.burelli@arduino.cc>